fix: minor code changes
