+++ src/components/common/SideContent.js
@@ -15,7 +15,7 @@ import Chatbox from './Chatbox';
 import { HEADER_HEIGHT, DRAWER_WIDTH } from '../../config/constants';
 import { LayoutContext } from '../context/LayoutContext';
 import Item from './Item';
-import ItemHeader from './ItemHeader'
+import BuilderButton from './BuilderButton'
 import {
   ITEM_CHATBOX_BUTTON_ID,
   PANNEL_CLOSE_BUTTON_ID,
@@ -228,7 +228,7 @@ const SideContent = ({ children, item }) => {
 
         {displayPinButton()}
 
-        <ItemHeader id={item.get('id')} />
+        <BuilderButton id={item.get('id')} />
 
         {children}
       </main>
