+++ src/components/common/Tree/CustomContentTree.js
@@ -1,0 +1,80 @@
+/* eslint-disable react/forbid-prop-types */
+import React from 'react';
+import PropTypes from 'prop-types';
+import clsx from 'clsx';
+import Typography from '@material-ui/core/Typography';
+import { useTreeItem } from '@mui/lab/TreeItem';
+
+const CustomContentTree = React.forwardRef((props, ref) => {
+    const {
+      classes,
+      className,
+      label,
+      nodeId,
+      icon: iconProp,
+      expansionIcon,
+      displayIcon,
+    } = props;
+  
+    const {
+      disabled,
+      expanded,
+      selected,
+      focused,
+      handleExpansion,
+      handleSelection,
+      preventSelection,
+    } = useTreeItem(nodeId);
+  
+    const icon = iconProp || expansionIcon || displayIcon;
+  
+    const handleMouseDown = (event) => {
+      preventSelection(event);
+    };
+  
+    const handleExpansionClick = (event) => {
+      handleExpansion(event);
+    };
+  
+    const handleSelectionClick = (event) => {
+      handleSelection(event);
+    };
+  
+    return (
+      // eslint-disable-next-line jsx-a11y/no-static-element-interactions
+      <div
+        className={clsx(className, classes.root, {
+          [classes.expanded]: expanded,
+          [classes.selected]: selected,
+          [classes.focused]: focused,
+          [classes.disabled]: disabled,
+        })}
+        onMouseDown={handleMouseDown}
+        ref={ref}
+      >
+        {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions */}
+        <div onClick={handleExpansionClick} className={classes.iconContainer}>
+          {icon}
+        </div>
+        <Typography
+          onClick={handleSelectionClick}
+          component="div"
+          className={classes.label}
+        >
+          {label}
+        </Typography>
+      </div>
+    );
+  });
+  
+  CustomContentTree.propTypes = {
+    classes: PropTypes.object.isRequired,
+    className: PropTypes.string.isRequired,
+    displayIcon: PropTypes.node.isRequired,
+    expansionIcon: PropTypes.node.isRequired,
+    icon: PropTypes.node.isRequired,
+    label: PropTypes.node.isRequired,
+    nodeId: PropTypes.string.isRequired,
+  };
+
+  export default CustomContentTree;
